+++ Reservation_Hotel/Reservation.h
@@ -21,5 +21,5 @@ public:
     int getMontantTotal() const;
 
     void modifierSejour(Date nouvelleDate, int nouvellesNuits);
-    int calculerMontantTotal(int nbNuits, double prix);
+    int calculerMontantTotal(int nbNuits, int prix);
 };
